arch/tools/pyproject.toml
[tool.poetry]
name = "archgw"
version = "0.1.3"
description = "Python-based CLI tool to manage Arch Gateway."
authors = ["Katanemo Labs, Inc."]
packages = [
    { include = "cli" }
]
readme = "README.md"

include = [
# Include package data (docker-compose.yaml and other files)[
    "config/docker-compose.yaml",
    "config/arch_config_schema.yaml",
]

[tool.poetry.dependencies]
python = ">=3.12"
archgw_modelserver = "0.1.3"
pyyaml = "^6.0.2"
pydantic = "^2.9.2"
click = "^8.1.7"
jinja2 = "^3.1.4"
jsonschema = "^4.23.0"
setuptools = "75.5.0"
huggingface_hub = "^0.26.0"

[tool.poetry.scripts]
archgw = "cli.main:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
